zmj 教学研究

+++ src/main/java/org/jplus/controller/teachingresearchController.java
@@ -4,6 +4,7 @@ import org.jplus.interceptor.NeedLogin;
 import org.jplus.pojo.Users;
 import org.jplus.pojo.teachingresearchInfo.*;
 import org.jplus.service.JxcgServiceImpl;
+import org.jplus.service.TjztService;
 import org.jplus.utils.GetWorkCount;
 import org.jplus.utils.GetYear;
 import org.jplus.utils.UserContext;
@@ -27,6 +28,9 @@ public class teachingresearchController {
 
     @Autowired
     JxcgServiceImpl jxcgService;
+
+    @Autowired
+    TjztService tjztService;
 
     //数据显示
     @NeedLogin
@@ -59,33 +63,44 @@ public class teachingresearchController {
     //删除教学成果某一列
     @NeedLogin
     @RequestMapping("/delJxcg")
-    public String delJxcg(Integer id) {
-        jxcgService.deleteJxcg(id);
-        return "redirect:/teachingresearch";
+    public String delJxcg(Integer id,Users user) {
+    //提交状态处于0，表示未提交可以操作数据
+        if (tjztService.getTjzt(user.getGh()).getTjzt() == 0) {
+            jxcgService.deleteJxcg(id);
+        }
+            return "redirect:/teachingresearch";
     }
 
     //添加教学成果某一列
     @NeedLogin
     @PostMapping("/addJxcg")
-    public String addJxcg(@ModelAttribute(value = "jxcg") Jxcg jxcg) {
-        //获取工号
-        jxcg.setGh(UserContext.getUser().getGh());
-        //获取工作量
-        jxcg.setJxcggzl(GetWorkCount.GetWorkCount(jxcg.getJxcgzrs(), jxcg.getXmpm(),
-                jxcgService.getGdgzl(jxcg.getCgdjbm())));
-        //存入年份
-        jxcg.setNd(GetYear.getYears());
-        //添加信息
-        jxcgService.addJxcg(jxcg);
-        return "redirect:/teachingresearch";
+
+    public String addJxcg(@ModelAttribute(value = "jxcg") Jxcg jxcg,Users user) {
+        //提交状态处于0，表示未提交可以操作数据
+        if (tjztService.getTjzt(user.getGh()).getTjzt() == 0) {
+            //获取工号
+            jxcg.setGh(user.getGh());
+            //获取工作量
+            jxcg.setGzl(GetWorkCount.GetWorkCount(jxcg.getZrs(), jxcg.getXmpm(),
+                    jxcgService.getGdgzl(jxcg.getCgdjbm())));
+            //存入年份
+            jxcg.setNd(GetYear.getYears());
+            //添加信息
+            jxcgService.addJxcg(jxcg);
+        }
+            return "redirect:/teachingresearch";
     }
 
     //删除纵向教研项目的某一列
     @NeedLogin
     @RequestMapping("/delZxxm")
-    public String delZxxm(Integer id) {
-        jxcgService.deleteZxxm(id);
+    public String delZxxm(Integer id,Users user) {
+        //提交状态处于0，表示未提交可以操作数据
+        if (tjztService.getTjzt(user.getGh()).getTjzt() == 0) {
+            jxcgService.deleteZxxm(id);
+        }
         return "redirect:/teachingresearch";
+
     }
 
     //添加纵向教研项目的某一列
@@ -95,7 +110,7 @@ public class teachingresearchController {
         //获取工号
         zxxm.setGh(user.getGh());
         //获取工作量
-        zxxm.setZxxmgzl(GetWorkCount.GetWorkCount(zxxm.getXmzrs(), zxxm.getXmpm(),
+        zxxm.setGzl(GetWorkCount.GetWorkCount(zxxm.getZrs(), zxxm.getXmpm(),
                 jxcgService.getZxxmgzl(zxxm.getXmdjbm())));
         //存入年份
         zxxm.setNd(GetYear.getYears());
@@ -108,7 +123,7 @@ public class teachingresearchController {
     //删除纵向教研项目的某一列
     @NeedLogin
     @RequestMapping("/delHxxm")
-    public String delHxxm(Integer id) {
+    public String delHxxm(Integer id,Users user) {
         jxcgService.deleteHxxm(id);
         return "redirect:/teachingresearch";
     }
@@ -120,7 +135,7 @@ public class teachingresearchController {
         //获取工号
         hxxm.setGh(user.getGh());
         //获取工作量
-        hxxm.setHxgzl((GetWorkCount.GetHxWorkCount(hxxm.getDzjf())>200) ? 200:
+        hxxm.setGzl((GetWorkCount.GetHxWorkCount(hxxm.getDzjf())>200) ? 200:
                 GetWorkCount.GetHxWorkCount(hxxm.getDzjf()));
         //存入年份
         hxxm.setNd(GetYear.getYears());
@@ -144,7 +159,7 @@ public class teachingresearchController {
         //获取工号
         jc.setGh(user.getGh());
         //获取工作量
-        jc.setJcgzl(GetWorkCount.GetWorkCount(jc.getXmzrs(), jc.getXmpm(),
+        jc.setGzl(GetWorkCount.GetWorkCount(jc.getZrs(), jc.getXmpm(),
                 jxcgService.getZxxmgzl(jc.getJcjbbm())));
         //存入年份
         jc.setNd(GetYear.getYears());
@@ -156,7 +171,7 @@ public class teachingresearchController {
     //删除教学改革某一列
     @NeedLogin
     @RequestMapping("/delJxgg")
-    public String delJxgg(Integer id) {
+    public String delJxgg(Integer id,Users user) {
         jxcgService.deleteJxgg(id);
         return "redirect:/teachingresearch";
     }
@@ -168,7 +183,7 @@ public class teachingresearchController {
         //获取工号
         jxgg.setGh(user.getGh());
         //获取工作量
-        jxgg.setJxgggzl(GetWorkCount.GetWorkCount(jxgg.getJxggzrs(), jxgg.getXmpm(),
+        jxgg.setGzl(GetWorkCount.GetWorkCount(jxgg.getZrs(), jxgg.getXmpm(),
                 jxcgService.getZxxmgzl(jxgg.getLxjbbm())));
         //存入年份
         jxgg.setNd(GetYear.getYears());
